Update App.js
In this lesson I am going to use localStorage to allow the notes to be saved into local storage.  I have to use JSON.parse and stringify allow the machine go back and forth as a string.  I am using useState to getItem and I am useEffect to setItem.  notes is the prop that we are altering.