Adding feature to updateNote function
in this update, I am creating a newArray variable.   With a for loop it iterates over the note in the oldNotes array.  For each note, it checks if the 'id' of the note matches the 'currentNoteId' and if they match it creates a new note object with body updated to the new text value and adds it to the beginning of newArray using unshift.

If the id doesn't match the currentNoteId, it pushes the old note as it is to the new array.

Finnally it returns the newArray which will contain the updated notes array with the note that mattches the currentNoteId.